Add simple i18n handling to controllers
This allows users of this package to extract their translations via
`seatbelt extract`, and kind-of merge them with `goi18n merge`. It's not
pretty, but it works for now.

In the future it'd be preferable to switch to gotext for translations,
but it's still a bit early to be using it, as their extraction tools are
quite rigid and conflict with the file location requirements that
Seatbelt currently has. That being said, gotext has much better support
for position based arguments that affect pluralization, translation
comments, source locations, etc, so the long term goal is to use that
package. The Go Authors are rumoured to be working on adding i18n to the
Go blog, and I'm assuming they'll use that package, so the gotext
package should hopefully be receiving some updates soon to make it
easier to work with HTML.