all: version public assets, improve sessions
Adds the ability to version public assets (JS, CSS, images, etc) in
templates by using the new `versionpath` helper. This appends a query
parameter containing the last modified time, in order to purge the
browser cache when assets are updated. This also required a change to
the way helper funcs are loaded, as there was previously a bug that
prevented user-defined template funcs from being used in layout files.

This also added some configuration options to the session cookie. The
name and max age of the session cookie can now be customized.