values: refactor into pkg, use ctx for storage
Updates the "values" API so that it used the request context for
storage, rather than relying on a per-request map stored in the context
instance.

This allows us to create arbitrary "values" instances from requests,
which makes it possible to interact with this API anytime we have an
HTTP request. This was required to merge the request-scoped "values" in
the I18N template helper, which was (in my opinion) required to make it
consistent with how the HTML rendering API and its usage of "values"
feels.